requirements.txt
Pillow
opencv-python
numpy
playsound3